+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Subtract `PEPOCH` from all TOA and `MJDParameter` values; add `epoch` attribute to `TimingModel`
 - Moved `basis_dot` to `jump.jl`
 - Better priors for `KOM`, `SINI`, `KIN`, `STIGMA`, `SHAPMAX`, `EFAC`, etc.
+- Rearranged code in `pyvela`
 ## Fixed
 - Unnecessary repetition of `sin` and `cos` in ecliptic coordinate conversion
 - Bug in `_ecorr_lnlike_group`